moved the test case into TestAdhocCreateDropView
